LayoutTests/inspector/cookie-parser-expected.txt

Tests inspector cookie parser

source: cookie=value
name: cookie, value: value, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 12
{
}
source: $version=1; a=b,c  =   d, e=f
name: a, value: b,c  =   d, e=f, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 17
{
}
source: $version=1; a=b;c  =   d; e =f
name: a, value: b, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 4
{
}
name: c, value: d, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 10
{
}
name: e, value: f, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 4
{
}
source: cooke1 = value1; another cookie = another value
name: cooke1, value: value1, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 25
{
}
name: cookie, value: another value, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 22
{
}
source: cooke1 = value; $Path=/; $Domain=.example.com;
name: cooke1, value: value, httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 46
{
    domain : ".example.com"
    path : "/"
}
source: cooke1 = value; $Path=/; $Domain=.example.com ; Cookie2 = value2; $Path = /foo; $DOMAIN = foo.example.com;
name: cooke1, value: value, httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 48
{
    domain : ".example.com"
    path : "/"
}
name: Cookie2, value: value2, httpOnly: false, secure: false, session: true, path: /foo, domain: foo.example.com, port: undefined, expires: n/a, size: 58
{
    domain : "foo.example.com"
    path : "/foo"
}
source: cooke1 = value; $Path=/; $Domain=.example.com
Cookie2 = value2; $Path = /foo; $DOMAIN = foo.example.com; 
name: cooke1, value: value, httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 46
{
    domain : ".example.com"
    path : "/"
}
name: Cookie2, value: value2, httpOnly: false, secure: false, session: true, path: /foo, domain: foo.example.com, port: undefined, expires: n/a, size: 59
{
    domain : "foo.example.com"
    path : "/foo"
}
source: $version =1; cooke1 = value; $Path=/; $Domain   =.example.com;  
 Cookie2 = value2; $Path = /foo; $DOMAIN = foo.example.com;
name: cooke1, value: value, httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 53
{
    domain : ".example.com"
    path : "/"
}
name: Cookie2, value: value2, httpOnly: false, secure: false, session: true, path: /foo, domain: foo.example.com, port: undefined, expires: n/a, size: 58
{
    domain : "foo.example.com"
    path : "/foo"
}
source: cookie=value
name: cookie, value: value, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 12
{
}
source: a=b
 c=d
 f
name: a, value: b, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 5
{
}
name: c, value: d, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 5
{
}
name: , value: f, httpOnly: false, secure: false, session: true, path: undefined, domain: undefined, port: undefined, expires: n/a, size: 1
{
}
source: cooke1 = value; Path=/; Domain=.example.com;
name: cooke1, value: value, httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 44
{
    domain : ".example.com"
    path : "/"
}
source: cooke1 = value; Path=/; Domain=  .example.com 
Cookie2 = value2; Path = /foo; Domain = foo.example.com
name: cooke1, value: value, httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 47
{
    domain : ".example.com"
    path : "/"
}
name: Cookie2, value: value2, httpOnly: false, secure: false, session: true, path: /foo, domain: foo.example.com, port: undefined, expires: n/a, size: 55
{
    domain : "foo.example.com"
    path : "/foo"
}
source: cooke1 = value; expires = Mon, Oct 18 2010 17:00 GMT+0000; Domain   =.example.com
Cookie2 = value2; Path = /foo; DOMAIN = foo.example.com; HttpOnly; Secure; Discard;
name: cooke1, value: value, httpOnly: false, secure: false, session: false, path: undefined, domain: .example.com, port: undefined, expires: 1287421200000, size: 82
{
    domain : ".example.com"
    expires : "Mon, Oct 18 2010 17:00 GMT+0000"
}
name: Cookie2, value: value2, httpOnly: true, secure: true, session: true, path: /foo, domain: foo.example.com, port: undefined, expires: n/a, size: 83
{
    discard : undefined
    domain : "foo.example.com"
    httponly : undefined
    path : "/foo"
    secure : undefined
}
source: cooke1 = value; max-age= 1440; Domain   =.example.com
 Cookie2 = value2; Path = /foo; DOMAIN = foo.example.com; HttpOnly; Secure; Discard;
name: cooke1, value: value, httpOnly: false, secure: false, session: false, path: undefined, domain: .example.com, port: undefined, expires: 1287422640000, size: 55
{
    domain : ".example.com"
    max-age : "1440"
}
name: Cookie2, value: value2, httpOnly: true, secure: true, session: true, path: /foo, domain: foo.example.com, port: undefined, expires: n/a, size: 83
{
    discard : undefined
    domain : "foo.example.com"
    httponly : undefined
    path : "/foo"
    secure : undefined
}
source: cooke1; Path=/; Domain=.example.com;
name: , value: cooke1, httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 36
{
    domain : ".example.com"
    path : "/"
}
source: cooke1=; Path=/; Domain=.example.com;
name: cooke1, value: , httpOnly: false, secure: false, session: true, path: /, domain: .example.com, port: undefined, expires: n/a, size: 37
{
    domain : ".example.com"
    path : "/"
}
